backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "duckdb>=1.0.0",
    "pandas>=2.0.0",
    "pydantic>=2.0.0",
]
